+++ CHANGELOG.md
@@ -8,6 +8,10 @@
 - [beta] Tweak the user guide.
 
 
+### Fixes
+
+- [beta] Fix an issue where importing legacy YAML syntax definitions from the Format settings could fail.
+
 
 7.0.0-rc (809)
 --------------------------